INT-2546 Move method PollerParser.configureTransactionAttributes to IntegrationNamespaceUtils
* Move method *PollerParser.configureTransactionAttributes* to *IntegrationNamespaceUtils*
* Remove method *configureTransactionAttributes* from class *JpaParserUtils*

For reference see: https://jira.springsource.org/browse/INT-2546
